Revert "journal"
