+++ public/index.html
@@ -2,10 +2,10 @@
 <html>
 <head>
 	<title>Login Page</title>
-  <link rel="stylesheet" href="style.css">
-  <script src="https://cdn.tailwindcss.com"></script>
-  <link href="/dist/output.css" rel="stylesheet">    
-  <script src="initialize.js"></script>	
+    <link rel="stylesheet" href="style.css">
+    <script src="https://cdn.tailwindcss.com"></script>
+    <link href="/dist/output.css" rel="stylesheet">
+    <script src="/firebase.js"></script>	
 	<script src="https://www.gstatic.com/firebasejs/8.2.5/firebase-app.js"></script>
 	<script src="https://www.gstatic.com/firebasejs/8.2.5/firebase-auth.js"></script>
 	<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
